// src/types/tagMod.type.ts
export type TagModType = {
	[index: string]: string | number | undefined
	date?: string // Modifiable
	album?: string // Modifiable
	album_artist?: string // Modifiable
	artist?: string // Modifiable
	composer?: string // Modifiable
	genre?: string // Modifiable
	title?: string // Modifiable
	track?: number // Modifiable
	rating?: number // Modifiable
	comment?: string // Modifiable
	disc?: number // Modifiable
}